Prevent build abort in mpi portion of overview collector build
On the Cray platform, mpicc is not available for the MPICH
mpi implementation.  This modification prevents the overview
mpi collector from aborting on build.  This step one in the
fix - preventing an abort.  Next step is to build using gcc
and using the includes and linking in the mpi libraries from
the mpi implementation like we do in the base mpi collectors.