Allow the overview mpi openmpi wrappers using a non-standard openmpi installation
Add code to allow the overview mpi openmpi wrappers to build with the openmpi
installation in /usr/lib64/openmpi.  Currently the code can not find mpicc,
which is required.  This code adds more paths for the cmake/FindOpenMPI
to find the installation and adds code in the collector to use the mpicc
found in the non-standard installation location."